try several different memory profiling techniques
heaptrack was causing a stack overflow in my rust main, seemed like an issue
with pyo3 because it was happening in ForceField::load, but couldn't debug. perf
mem doesn't work on AMD, so valgrind was the best bet. showed no leaks, but I
didn't let it run for long because it was insanely slow